# app/api/api-keys/[id]/route.ts
import { NextRequest, NextResponse } from 'next/server'

import {
  applyAuthCookies,
  authenticatedFetch,
  getApiBaseUrl,
  hasAuthSession,
} from '@/app/api/_lib/controlPlane'

const API_BASE_URL = getApiBaseUrl()

export const dynamic = 'force-dynamic'

export async function DELETE(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  try {
    if (!hasAuthSession(request)) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
    }

    const { id } = await params

    const { response, tokenRefreshed, refreshedTokens } = await authenticatedFetch(
      request,
      `${API_BASE_URL}/v1/api-keys/${id}`,
      {
        method: 'DELETE',
        cache: 'no-store',
      }
    )

    if (!response.ok) {
      const payload = await response.json().catch(() => ({ detail: 'Failed to revoke API key' }))
      const nextResponse = NextResponse.json(payload, { status: response.status })

      if (tokenRefreshed && refreshedTokens) {
        applyAuthCookies(nextResponse, request, refreshedTokens)
      }

      return nextResponse
    }

    const nextResponse = new NextResponse(null, { status: 204 })

    if (tokenRefreshed && refreshedTokens) {
      applyAuthCookies(nextResponse, request, refreshedTokens)
    }

    return nextResponse
  } catch (error) {
    console.error('API key delete error:', error)
    return NextResponse.json({ error: 'Failed to connect to API' }, { status: 500 })
  }
}